Restructure code and styling
It was requested to take a model-based approached, not the data-centric
approach used at first, for organizing the code. This encourages a
separation of the concerns of ontologists from designers.

For example, the ontologist models a Person and its attributes in
entity.js. The designer can use person.html for displaying a profile and
reuse those models in people.html for displaying a list of people.